+++ ParaViewCore/VTKExtensions/Default/vtkPVAMRDualClip.cxx
@@ -67,8 +67,8 @@ int vtkPVAMRDualClip::RequestData(vtkInformation* vtkNotUsed(request),
   this->IsoValue = (this->VolumeFractionSurfaceValue *
                     PV_AMR_SURFACE_VALUE_UNSIGNED_CHAR);
 
-  size_t noOfArrays = this->Implementation->CellArrays.size();
-  for(size_t i=0; i < noOfArrays; ++i)
+  unsigned int noOfArrays = static_cast<unsigned int>(this->Implementation->CellArrays.size());
+  for(unsigned int i=0; i < noOfArrays; ++i)
     {
     vtkMultiBlockDataSet* out = this->DoRequestData(
       hbdsInput, this->Implementation->CellArrays[i].c_str());
